+++ processor/src/test/fixtures/input/com/example/basic/DeprecatedReactComponent.java
@@ -1,0 +1,17 @@
+package com.example.basic;
+
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+
+@Deprecated
+@ReactComponent
+abstract class DeprecatedReactComponent
+  extends Component
+{
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}