+++ processor/src/test/fixtures/input/com/example/post_mount/DeprecatedPostMountModel.java
@@ -1,0 +1,23 @@
+package com.example.post_mount;
+
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.PostMount;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class DeprecatedPostMountModel
+  extends Component
+{
+  @Deprecated
+  @PostMount
+  void postMount()
+  {
+  }
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}